remove extra blank lines in chapter2
And if none of those are set then it uses vi on your system.
If you are not used to it, vi is a confusing and complicated experience that can be off-putting.
Good alternatives to vi for composing your commit messages are nano,
pico and micro depending on what your system has installed.
Nano includes shortcut help at the bottom of the display to help you figure out how to save and exit.

Well configured command line editors will provide you with syntax highlighting and automatic line
wrapping to help you keep to the appropriate structure of a commit message without confining you to it.

It is possible to use non-command line based editors provided they can be launched in a "blocking"
fashion so that the git commit command can assume that the commit message will have been completed by
 the time your editor closes. An example of this is running gvim with the -f (foreground) flag.
 It is harder to use IDE-like editors in this role as they don't tend to be opened
 and closed on a per-file basis.
 Github has documentation for some popular editors.